refactoring: solution sorting moved to fw package
